+++ src/torch_pointcloud/metrics/classification.py
@@ -5,7 +5,7 @@ from typing import Dict, Literal, Optional, Sequence, Tuple, Union, overload
 import torch
 from torch import Tensor
 
-from torch_pointcloud.ops.utils import safe_divide
+from torch_pointcloud.ops.math import safe_divide
 
 
 def confusion_matrix(